Aborting mission FlxNapeSprite
I decided to give up on making FlxNapeTilemap work with tilemaps imported from Ogmo for the sake of continuing the game.

I reverted from Nape based movement to a movement that uses flixel's speed, I kept the version of the project with the basic human class and nape movement in a separate repo (to be created) so that I can continue working on it in the future.

I succesfully imported a rough first city and made collisions work.